Replaced AttendanceEvent by TimeSlot
