+++ plan/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 from django.urls import reverse
 from django.utils.html import format_html
 
+from plan.admin_actions import update_attendees
 from plan.models import TimeSlot, SlotType, Event, Booking
 from django.utils.translation import ugettext as _
 
@@ -14,7 +15,7 @@ class SlotTypeAdmin(admin.ModelAdmin):
 @admin.register(TimeSlot)
 class TimeSlotAdmin(admin.ModelAdmin):
     list_display = ['name', 'begin', 'end', 'relevant_for', 'link_attendance_list']
-    #actions = [update_attendees]
+    actions = [update_attendees]
 
     @staticmethod
     def link_attendance_list(event):
