+++ plan/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
 from ophasebase.models import Ophase, Room, OphaseCategory
-from staff.models import StaffFilterGroup, TutorGroup
+from staff.models import TutorGroup
 
 
 class SlotType(models.Model):
@@ -30,7 +30,7 @@ class TimeSlot(models.Model):
     begin = models.DateTimeField(verbose_name=_("Beginn"))
     end = models.DateTimeField(verbose_name=_("Ende"))
     category = models.ManyToManyField(OphaseCategory)
-    relevant_for = models.ForeignKey(StaffFilterGroup, verbose_name=_("Filterkriterium: Wer könnte anwesend sein?"), null=True, on_delete=models.SET_NULL)
+    relevant_for = models.ForeignKey("staff.StaffFilterGroup", verbose_name=_("Filterkriterium: Wer könnte anwesend sein?"), null=True, on_delete=models.SET_NULL)
     attendance_required = models.BooleanField(blank=True, default=False)
     ophase = models.ForeignKey(Ophase, models.CASCADE)
     public = models.BooleanField(blank=True, default=False)
