+++ staff/admin.py
@@ -14,7 +14,7 @@ from .admin_actions import (
     staff_nametag_export,
     staff_overview_export,
     tutorgroup_export,
-    update_attendees, mark_attendance_x, mark_attendance_a, mark_attendance_e, mark_phoned_x, mark_phoned_e,
+    mark_attendance_x, mark_attendance_a, mark_attendance_e, mark_phoned_x, mark_phoned_e,
     mark_phoned_n, mark_attendance_v, generate_orga_cert)
 from .models import (
     HelperJob,
@@ -23,8 +23,7 @@ from .models import (
     Settings,
     TutorGroup,
     StaffFilterGroup,
-    Attendance,
-    AttendanceEvent
+    Attendance
 )
 
 
@@ -134,19 +133,6 @@ class AttendanceAdmin(admin.ModelAdmin):
     @staticmethod
     def person_phone(event):
         return event.person.phone
-
-
-@admin.register(AttendanceEvent)
-class AttendanceAdmin(admin.ModelAdmin):
-    list_display = ['name', 'begin', 'end', 'required_for', 'link_attendance_list']
-    actions = [update_attendees]
-
-    @staticmethod
-    def link_attendance_list(event):
-        return format_html('<a href="{url}?event__id__exact={id}">{name}</a>',
-                           url=reverse('admin:staff_attendance_changelist'),
-                           id=event.pk,
-                           name=_('Teilnehmerliste'))
 
 
 @admin.register(Settings)
